Register rest-client-reactive response for reflection
Fixes: #20200
