+++ integration-tests/rest-client-reactive/src/main/java/io/quarkus/it/rest/client/main/Apple.java
@@ -1,9 +1,5 @@
 package io.quarkus.it.rest.client.main;
 
-import io.quarkus.runtime.annotations.RegisterForReflection;
-
-// TODO get rid of need for this
-@RegisterForReflection
 public class Apple {
     private String cultivar;
